pyproject.toml
[project]
name = "video-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2.6",
    "django-filter>=25.1",
    "djangorestframework>=3.16.1",
    "djoser>=2.3.3",
    "drf-yasg>=1.21.10",
    "gunicorn>=23.0.0",
    "psycopg>=3.2.9",
    "python-dotenv>=1.1.1",
]

[dependency-groups]
dev = [
    "django-debug-toolbar>=6.0.0",
    "faker>=37.6.0",
]
